Fix cast

+++ flox/aggregate_npg.py
@@ -113,7 +113,7 @@ def _var_std_wrapper(group_idx, array, engine, *, axis=-1, **kwargs):
     # Attempt to increase numerical stability by subtracting the first element.
     # https://en.wikipedia.org/wiki/Algorithms_for_calculating_variance
     # Cast any unsigned types first
-    dtype = np.result_type(array, -1 * array[0])
+    dtype = np.result_type(array, np.int8(-1) * array[0])
     array = array.astype(dtype)
     first = _get_aggregate(engine).aggregate(group_idx, array, func="nanfirst", axis=axis)
     array = array - first[..., group_idx]
